move getUser to databaseClient
